Murder the competition cabling budget
The sr2013/competition/cabling budget line was for audio and power
cabling, both of which are veing provided by the SUSU and are now
addressed in the competition/venue-hire. All that remains is the cost
of purchasing RCDs.

This reduces the budget line down to just the cost of RCD purchase,
with the side-effect of making it non-consumable, and renames it.