Assign some funds to venue hire
We have £12k coming in from the Motorola Solutions Foundation towards
the hire of the venue.  The venue will cost £17580 in total, so we
need to allocate more money to it.

This patch allocates £5580 of our current free budget towards covering
this cost.  When we receive the latest batch of funds from the
Motorola Solutions Foundation, this will combine to meet the total
cost.

The deposit on the venue will be 25% of the total fee, which is
£4395.  Therefore, this partial assignment will cover the deposit.