Refuse to generate merges on the budget
Consider the following case:

 (1) The HEAD of the budget is commit A, which has £500 unallocated.
 (2) Commit B, with commit A as its head, is posted to code review. It
     allocates the £500 to purchasing fancy beehives.
 (3) Commit C, also with commit A as its head, is posted to code review. It
     allocates the £500 to purchasing fleen rotivators.
 (4) The SC verify that B has no overspend (which it does not), approve,
     and submit it.
 (5) The SC verify that C has no overspend (Which it does not), approve,
     and submit it.
 (6) Gerrit merges the two together, leaving budget HEAD with a £500
     overspend.

The issue here is that gerrit can create merges which leave the budget in
a state which is unreviewed.

This commit switches the submit action to "fast forward only" which will cause
Gerrit not to create merges. In the scenario above, step (5) would result
in C's rejection and need for rebasing onto B, after which reverification
would fail due to the overspend.