Add budget for SR2015 kit and kickstart
This adds budget for SR2015 kit manufacture and SR2015 kickstart.

In order to meet our kickstart deadline, the purchases relating to this
need to happen very soon.

This commit is the result of the sr2015-kit branch that I was
maintaining for a while.  This is that branch all squashed into one.

Please note that in order to achieve this, we had to introduce a new
feature into the budgeting.  Budget entries now have an optional
"uncertainty" entry.  If this is not present, then the uncertainty
defaults to 10%, which is the "fudge factor" that we previously
applied.  This allows us to deal with the situation that we are
currently in: we are going to be buying things in a few days, and we are
looking at the quotes for the things that we need to buy, so we know
with much more precision than normal how much they will cost.  This
means that you will likely need to update your tools.git in order for
budget-check to report that this is OK.

Change-Id: Id6c1d4ea12a6a20800d4f908fb89418d7e99398b